pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "error-log-agent-v2"
version = "2.0.0"
description = "Error log monitoring agent with K3s deployment pipeline"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    # LangGraph / LangChain
    "langgraph>=0.4.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    # Slack
    "slack-bolt[async]>=1.22.0",
    "slack-sdk>=3.35.0",
    "aiohttp>=3.10.0",
    # Database
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    # HTTP
    "httpx>=0.28.0",
    # Search
    "tavily-python>=0.5.0",
    # Config
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0",
    # Logging
    "structlog>=24.4.0",
    # Scheduler
    "apscheduler>=3.10.0",
    # Git
    "gitpython>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
